+++ test/dynamo/test_decorators.py
@@ -215,7 +215,7 @@ class DecoratorTests(torch._dynamo.test_case.TestCase):
             yield
 
         with mock.patch(
-            "torch._guards.TracingContext.current_frame",
+            "torch._tracing_context.TracingContext.current_frame",
             side_effect=global_context_capture_fn,
         ):
             torch._dynamo.optimize("eager")(e)(x)
