+++ torch/_dynamo/symbolic_convert.py
@@ -22,7 +22,8 @@ from unittest.mock import patch
 
 import torch
 import torch._logging
-from torch._guards import Checkpointable, tracing, TracingContext
+from torch._guards import Checkpointable
+from torch._tracing_context import tracing, TracingContext
 
 from . import (
     config,
