+++ torch/_dynamo/variables/builder.py
@@ -20,10 +20,11 @@ except ModuleNotFoundError:
 import torch
 
 from torch import SymInt
-from torch._guards import GuardSource, TracingContext
+from torch._guards import GuardSource
 from torch._ops import HigherOrderOperator
 from torch._streambase import _EventBase, _StreamBase
 from torch._subclasses.fake_tensor import FakeTensor, is_fake, maybe_get_fake_mode
+from torch._tracing_context import TracingContext
 from torch.fx.experimental.symbolic_shapes import (
     _constrain_range_for_size,
     CreateSymbolicPolicy,
