+++ torch/_guards.py
@@ -1,17 +1,12 @@
 from __future__ import annotations
 
-import contextlib
-
 import dataclasses
 import enum
 import functools
 import logging
 import threading
-import traceback
-import unittest.mock
 import weakref
 from abc import ABC, abstractmethod
-from contextlib import contextmanager
 from typing import (
     Any,
     Callable,
@@ -27,7 +22,6 @@ from typing import (
 )
 
 import torch
-from torch.utils import _pytree as pytree
 from torch.utils._traceback import CapturedTraceback
 
 log = logging.getLogger(__name__)
@@ -47,32 +41,6 @@ torch._guards is the definitional source of truth for general purpose guard stru
 An important thing to keep in mind here is the preservation of layering. There should be no dynamo notions,
 and no guard installation notions here.
 """
-
-
-class CompileId(NamedTuple):
-    frame_id: int
-    # This id is per-frame, and counts how many times we've compiled this
-    # frame.  This could have been a global id but having this be per-frame
-    # gives you a better intuitive sense for how many recompiles have occurred
-    # so far.
-    frame_compile_id: int
-    # TODO: consider also tracking the recompilation count
-
-    def __str__(self):
-        return f"{self.frame_id}/{self.frame_compile_id}"
-
-
-class TraceId(NamedTuple):
-    compile_id: CompileId
-    # This starts off as 0, and every time we restart analysis it goes
-    # up by one
-    attempt: int
-
-    def __str__(self):
-        if self.attempt == 0:
-            return str(self.compile_id)
-        else:
-            return f"{self.compile_id}_{self.attempt}"
 
 
 class GuardSource(enum.Enum):
@@ -312,6 +280,8 @@ class DuplicateInputs(GuardEnvExpr):
         assert self.input_source_a != self.input_source_b
 
 
+# TODO(voz): Move all the checkpointable, contexts, context state, etc
+# into its own file
 """
 Checkpointable is an interface for driving state snapshotting, left purposely vague for now.
 
@@ -490,7 +460,7 @@ class GuardsSet:
         if guard.stack is None:
             guard.stack = CapturedTraceback.extract(skip=1 + skip)
         if guard.user_stack is None:
-            guard.user_stack = TracingContext.extract_stack()
+            guard.user_stack = torch._tracing_context.TracingContext.extract_stack()
         self.inner.add(guard)
 
     def update(self, *others: Set[Guard]):
@@ -511,251 +481,6 @@ class GuardsContext(Checkpointable[GuardsCheckpointState]):
         # NB: "steals" the passed in state
         assert isinstance(state, GuardsCheckpointState)
         self.dynamo_guards = GuardsSet(state.dynamo_guards)
-
-
-_TLS = threading.local()
-
-"""
-TracingContext is the source of truth for all currently accumulated information
-needed to trace. Its lifecycle is kept 1:1 when using TorchDynamo, but other systems
-are open to managing their own TracingContext with that in mind.
-
-The purpose of TracingContext is not to be a dumping ground, or god object, but rather to avoid
-having to plumb complex subsystems across multiple verticals.
-
-Ex: A common example is guard accumulation between dynamo, shape_env, aot_autograd, and inductor.
-Accessing the current tracing context via
-TracingContext.get() allows users to accumulate their own guards for processing, without needing to know how
-to plumb objects back up to where frame interpretation happened.
-
-Note that you can end up with multiple TracingContext for a single compilation
-of a frame, as we reset the TracingContext whenever we restart analysis.
-CompileContext is a more overarching context that encompasses multiple restarts.
-"""
-
-
-class CompileContext:
-    @staticmethod
-    def get() -> CompileContext:
-        assert _TLS.compile_context is not None
-        return _TLS.compile_context
-
-    @staticmethod
-    def try_get() -> Optional[CompileContext]:
-        return getattr(_TLS, "compile_context", None)
-
-    def __init__(self, compile_id):
-        assert compile_id is None or isinstance(compile_id, CompileId)
-        self.compile_id: Optional[CompileId] = compile_id
-        self.attempt = 0
-
-    @staticmethod
-    def current_compile_id():
-        self = CompileContext.try_get()
-        if self is None:
-            return None
-        return self.compile_id
-
-    @staticmethod
-    def current_trace_id():
-        self = CompileContext.try_get()
-        if self is None:
-            return None
-        if self.compile_id is None:
-            return None
-        return TraceId(self.compile_id, self.attempt)
-
-
-class TracingContext:
-    """
-    Provides the currently installed TracingContext, or None.
-
-    Note that it is a staticmethod, and invocations outside of `with tracing()` (see below), are valid but
-    will return None.
-    """
-
-    @staticmethod
-    def try_get() -> Optional[TracingContext]:
-        return getattr(_TLS, "tracing_context", None)
-
-    @staticmethod
-    def get() -> TracingContext:
-        if ctx := TracingContext.try_get():
-            return ctx
-        raise RuntimeError(
-            "TracingContext.get() must be called within an ongoing trace."
-        )
-
-    def __init__(self, fake_mode):
-        self.guards_context = GuardsContext()
-        self.module_context = ModuleContext()
-        self.global_context = GlobalContext()
-        self.fake_mode = fake_mode
-        self.frame_summary_stack = []
-        # This is morally part of frame_summary_stack, but it is kept separate
-        # for clarity.  As we process a frame, this variable gets updated
-        # to keep track of what line we are in the function.  We make a
-        # function call, this gets cleared and the frame location is pushed
-        # to frame_summary_stack (prepping this variable for the inner frame's
-        # progress)
-        self.loc_in_frame = None
-        # this is only set after aot_autograd
-        self.fw_metadata = None
-        self.params_flat = None
-        # this is for extended return calling convention from backend
-        # compiler to aot_autograd
-        # Per output, what the compiler specified stride of the output is,
-        # or None if no stride is known.  This is always the HINT, it
-        # is never a SymInt (it would be better if it was a SymInt, but
-        # I can't conveniently get this from Inductor atm.  Also, be
-        # careful not to accidentally induce guards on the SymInt if
-        # you ever do change this in aot_autograd.py; you should check
-        # on permutations preferentially.)
-        self.output_strides: Optional[List[Optional[List[int]]]] = None
-        # When this is True, whenever we encounter an int in Dynamo tracing,
-        # we will (1) force unspec it and (2) force it as a size-like unbacked
-        # integer.  This is currently used when processing certain lists of
-        # ints that are known to be size-like and may have 0/1 entries that we
-        # must not specialize on.
-        self.force_unspec_int_unbacked_size_like = False
-
-    @staticmethod
-    @contextmanager
-    def patch(**kwargs):
-        prior = {}
-        ctx = TracingContext.get()
-
-        for key in kwargs.keys():
-            # KeyError on invalid entry
-            prior[key] = getattr(ctx, key)
-        for key, val in kwargs.items():
-            setattr(ctx, key, val)
-        try:
-            yield
-        finally:
-            for key, val in prior.items():
-                setattr(ctx, key, val)
-
-    @staticmethod
-    def extract_stack():
-        self = TracingContext.try_get()
-        if self is None:
-            return traceback.StackSummary()
-        stack = list(self.frame_summary_stack)
-        if self.loc_in_frame is not None:
-            stack.append(self.loc_in_frame)
-        return traceback.StackSummary.from_list(stack)
-
-    # Call this when you want to call into some code that isn't necessarily
-    # associated with the current frame state
-    @staticmethod
-    @contextlib.contextmanager
-    def clear_frame():
-        tc = TracingContext.get()
-        with unittest.mock.patch.object(
-            tc, "frame_summary_stack", []
-        ), unittest.mock.patch.object(tc, "loc_in_frame", None):
-            try:
-                yield
-            except Exception as e:
-                # Prevent real_stack from getting attached
-                #
-                # The invariant is that if an Exception as real_stack, we've
-                # appropriately attached a user stack and we no longer need to
-                # attach anything. Because we cannot conveniently interpose
-                # when an exception is thrown, we instead interpose everywhere
-                # we set what the user stack is set (using the context
-                # manager). However, our compiler stack does "tail calls"
-                # (when it calls into user compiler), at which point the
-                # parent exception frames would incorrectly attach an
-                # incorrect frame.
-                #
-                # However, if, somehow, someone raised an exception with this
-                # scope that had a stack (for example, because they are
-                # restoring the user stack state appropriately as they process
-                # node by node), we should respect it. Thus, we cannot
-                # unconditionally set None.
-                if not hasattr(e, "real_stack"):
-                    e.real_stack = None  # type: ignore[attr-defined]
-                raise
-
-    @staticmethod
-    @contextlib.contextmanager
-    def current_frame(frame_summary):
-        # frame_summary can be None to solely take advantage of real_stack
-        # attachment to thrown exceptions
-        tc = TracingContext.get()
-        if frame_summary is not None:
-            tc.frame_summary_stack.append(frame_summary)
-        old = tc.loc_in_frame
-        tc.loc_in_frame = None
-        try:
-            yield
-        except Exception as e:
-            if not hasattr(e, "real_stack"):
-                e.real_stack = tc.extract_stack()  # type: ignore[attr-defined]
-            raise
-        finally:
-            if frame_summary is not None:
-                tc.frame_summary_stack.pop()
-            tc.loc_in_frame = old
-
-    @staticmethod
-    @contextlib.contextmanager
-    def report_output_strides():
-        tc = TracingContext.try_get()
-        if tc is None:
-            yield None
-            return
-        old_output_strides = tc.output_strides
-        tc.output_strides = []
-        try:
-            yield tc.output_strides
-        finally:
-            tc.output_strides = old_output_strides
-
-    @staticmethod
-    def set_current_loc(filename, lineno, frame_name):
-        TracingContext.get().loc_in_frame = traceback.FrameSummary(
-            filename, lineno, frame_name
-        )
-
-
-@contextmanager
-def compile_context(context: CompileContext):
-    old_context = getattr(_TLS, "compile_context", None)
-    _TLS.compile_context = context
-    try:
-        yield context
-    finally:
-        _TLS.compile_context = old_context
-
-
-@contextmanager
-def tracing(context: Optional[TracingContext]):
-    """
-    This function installs the passed in tracing context as a dynamic scoped
-    global variable.
-
-    Calls to TracingContext.get() while not under a `with tracing()` context
-    will return None.
-    """
-    old_context = getattr(_TLS, "tracing_context", None)
-    _TLS.tracing_context = context
-    try:
-        yield context
-    except Exception as e:
-        if not hasattr(e, "real_stack") and context is not None:
-            e.real_stack = context.extract_stack()  # type: ignore[attr-defined]
-        raise
-    finally:
-        if (
-            context is not None
-            and context.fake_mode is not None
-            and context.fake_mode.shape_env is not None
-        ):
-            context.fake_mode.shape_env.cleanup()
-        _TLS.tracing_context = old_context
 
 
 # Subclasses can be found in torch/_dynamo/source.py
@@ -784,47 +509,3 @@ class Source:
 @dataclasses.dataclass(frozen=True)
 class ChainedSource(Source):
     base: Source
-
-
-def detect_fake_mode(inputs: Any = None):
-    """
-    Attempts to "detect" what the current fake mode is.  If there is one ambiently
-    available from TracingContext, we preferentially use that.  Otherwise, we
-    heuristically detect the fake mode via the following sources, in order of
-    priority:
-
-        - Currently active fake mode on stack
-        - Fake mode associated with passed in tensors (inputs does not
-          have to be flattened)
-    """
-    from torch._subclasses.fake_tensor import FakeTensor, FakeTensorMode
-
-    fake_modes = []
-
-    if context := TracingContext.try_get():
-        fake_mode = context.fake_mode
-        if fake_mode is not None:
-            fake_modes.append((fake_mode, "tracing context", 0))
-
-    from torch.utils._python_dispatch import _get_current_dispatch_mode_stack
-
-    for i, m in enumerate(reversed(_get_current_dispatch_mode_stack())):
-        if isinstance(m, FakeTensorMode):
-            fake_modes.append((m, "active fake mode", i))
-
-    flat_inputs = pytree.tree_leaves(inputs)
-    for i, flat_input in enumerate(flat_inputs):
-        if isinstance(flat_input, FakeTensor):
-            fake_modes.append((flat_input.fake_mode, "fake tensor input", i))
-
-    if fake_modes:
-        fake_mode, desc1, i1 = fake_modes[0]
-        for m, desc2, i2 in fake_modes[1:]:
-            assert fake_mode is m, (
-                f"fake mode ({fake_mode}) from {desc1} {i1} doesn't match mode ({m}) from {desc2} {i2}\n\n"
-                f"fake mode from {desc1} {i1} allocated at:\n{fake_mode.stack}\n"
-                f"fake mode from {desc2} {i2} allocated at:\n{m.stack}"
-            )
-        return fake_mode
-    else:
-        return None
