+++ torch/_inductor/fx_passes/joint_graph.py
@@ -4,7 +4,7 @@ from collections import Counter
 from typing import Dict, Set
 
 import torch
-import torch._guards
+import torch._tracing_context
 from torch._inductor.constant_folding import ConstantFolder
 from torch.multiprocessing.reductions import StorageWeakRef
