+++ torch/_inductor/lowering.py
@@ -5056,7 +5056,7 @@ register_inplace(aten.__ixor__, aten.__xor__)
 
 @register_lowering(aten.sym_constrain_range)
 def sym_constrain_range(a, min, max):
-    tracing_context = torch._guards.TracingContext.get()
+    tracing_context = torch._tracing_context.TracingContext.get()
     assert a in tracing_context.fake_mode.shape_env.var_to_range
     return a
