+++ torch/fx/experimental/symbolic_shapes.py
@@ -33,7 +33,8 @@ from torch.fx.experimental.sym_node import SymNode, SymTypes
 
 # NB: The sym_* functions are used via getattr() and must be imported here.
 from torch import SymBool, SymFloat, SymInt
-from torch._guards import ShapeGuard, Source, TracingContext
+from torch._guards import ShapeGuard, Source
+from torch._tracing_context import TracingContext
 from torch.utils._python_dispatch import is_traceable_wrapper_subclass
 from torch.utils._sympy.functions import FloorDiv, Mod, IsNonOverlappingAndDenseIndicator
 from torch.utils._sympy.solve import try_solve
